Abstracted search options into a component
