+++ src/screens/Dashboard.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useAuth } from "react-oidc-context";
 
-import { useRecoilState } from "recoil";
+import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
 import {
   dataRouteAtom,
   dataStreamIdAtom,
@@ -9,9 +9,8 @@ import {
 } from "src/state/searchParams";
 import { dataStreamsAtom } from "src/state/dataStreams";
 
-import { Dropdown } from "@us-gov-cdc/cdc-react";
-import { Icons } from "@us-gov-cdc/cdc-react-icons";
 import PieChart from "src/components/PieChart";
+import SearchOptions from "src/components/SearchOptions";
 import StatusBoxes from "src/components/StatusBoxes";
 
 import getReportCounts, {
@@ -20,12 +19,8 @@ import getReportCounts, {
 } from "src/utils/api/reportCounts";
 
 import { getPastDate } from "src/utils/helperFunctions/date";
-import timeframes, { Timeframe } from "src/types/timeframes";
 import getDataStreams, { DataStream } from "src/utils/api/dataStreams";
-import {
-  getDataRoutes,
-  getDataStreamIds,
-} from "src/utils/helperFunctions/dataStreams";
+import { getDataRoutes } from "src/utils/helperFunctions/dataStreams";
 
 function Dashboard() {
   const auth = useAuth();
@@ -34,10 +29,10 @@ function Dashboard() {
 
   const [dataStream, setDataStream] = useRecoilState(dataStreamIdAtom);
   const [dataRoute, setDataRoute] = useRecoilState(dataRouteAtom);
-  const [timeframe, setTimeframe] = useRecoilState(timeFrameAtom);
+  const timeframe = useRecoilValue(timeFrameAtom);
 
   // TODO: Replace with global state
-  const [dataStreams, setDataStreams] = useRecoilState(dataStreamsAtom);
+  const setDataStreams = useSetRecoilState(dataStreamsAtom);
   useEffect(() => {
     const fetchCall = async () => {
       const res = await getDataStreams(auth.user?.access_token || "");
@@ -81,48 +76,10 @@ function Dashboard() {
     if (dataStream) fetchCall();
   }, [auth, dataStream, dataRoute, timeframe]);
 
-  const handleDataStream = (dataStreamId: string) => {
-    setDataStream(dataStreamId);
-    const route = getDataRoutes(dataStreams, dataStreamId)[0];
-    setDataRoute(route);
-  };
-
-  const handleTimeframe = (time: string) => {
-    const timeframe = time as Timeframe;
-    setTimeframe(timeframe);
-  };
-
   return (
     <section className="main_content padding-x-2">
       <h1 className="cdc-page-header padding-y-3 margin-0">Dashboard</h1>
-      <div className="padding-bottom-2 display-flex flex-row flex-justify">
-        <div className="display-flex flex-row cdc-submissions-page--filters">
-          <Dropdown
-            className="padding-right-2"
-            items={getDataStreamIds(dataStreams)}
-            label="Data Stream"
-            onSelect={handleDataStream}
-            srText="Data Stream"
-            defaultValue={dataStream}
-          />
-          <Dropdown
-            className="padding-right-2"
-            items={getDataRoutes(dataStreams, dataStream)}
-            label="Route"
-            onSelect={setDataRoute}
-            srText="Data Route"
-            defaultValue={dataRoute}
-          />
-          <Dropdown
-            items={timeframes}
-            label="Timeframe"
-            labelIcon={<Icons.Calendar />}
-            onSelect={handleTimeframe}
-            srText="Timeframe"
-            defaultValue={timeframe}
-          />
-        </div>
-      </div>
+      <SearchOptions />
       <StatusBoxes data={countsData} />
       <PieChart data={countsData} />
     </section>
